requirements.txt
streamlit>=1.31
anthropic>=0.34.0
python-dotenv>=1.0.1
pymysql>=1.1.1
openpyxl>=3.1.5
sqlglot>=25.0.0
pandas>=2.2.0
plotly>=5.24.0
SQLAlchemy>=2.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
